PizzaDelivery/pizza_delivery/requirements.txt
asgiref==3.6.0
Django==4.1.7
psycopg2==2.9.5
sqlparse==0.4.3
tzdata==2023.3
kafka-python
Werkzeug==2.2.3
pillow